move sync query support to new opt
