+++ policy/policy_handler/subscription.go
@@ -2,17 +2,13 @@ package policy_handler
 
 import (
 	"context"
-	"fmt"
-
 	"github.com/atomist-skills/go-skill"
-	"github.com/atomist-skills/go-skill/policy/data"
 	"olympos.io/encoding/edn"
 )
 
-func WithSubscription() Opt {
+func withSubscription() Opt {
 	return func(h *EventHandler) {
 		h.subscriptionDataProviders = append(h.subscriptionDataProviders, getSubscriptionData)
-		h.dataSourceProviders = append(h.dataSourceProviders, getSyncDataSources)
 	}
 }
 
@@ -22,15 +18,4 @@ func getSubscriptionData(ctx context.Context, req skill.RequestContext) ([][]edn
 	}
 
 	return req.Event.Context.Subscription.Result, req.Event.Context.Subscription.Configuration, nil
-}
-
-func getSyncDataSources(ctx context.Context, req skill.RequestContext) ([]data.DataSource, error) {
-	gqlDs, err := data.NewSyncGraphqlDataSource(ctx, req)
-	if err != nil {
-		return nil, fmt.Errorf("unable to create data source: %w", err)
-	}
-
-	return []data.DataSource{
-		gqlDs,
-	}, nil
 }
